feat!: huge settings view revamp for macOS 15+ users
